+++ src/ui/svelte/flyouts/state/mod.svelte.ts
@@ -2,24 +2,27 @@ import { invoke, SeelenCommand, SeelenEvent, subscribe } from "@seelen-ui/lib";
 import { lazyRune } from "libs/ui/svelte/utils";
 
 let mediaDevices = lazyRune(() => invoke(SeelenCommand.GetMediaDevices));
-await subscribe(SeelenEvent.MediaDevices, mediaDevices.setByPayload);
-await mediaDevices.init();
+subscribe(SeelenEvent.MediaDevices, mediaDevices.setByPayload);
 
 let mediaPlaying = lazyRune(() => invoke(SeelenCommand.GetMediaSessions));
-await subscribe(SeelenEvent.MediaSessions, mediaPlaying.setByPayload);
-await mediaPlaying.init();
+subscribe(SeelenEvent.MediaSessions, mediaPlaying.setByPayload);
 
 let brightness = lazyRune(() => invoke(SeelenCommand.GetAllMonitorsBrightness));
-await subscribe(SeelenEvent.SystemMonitorsBrightnessChanged, brightness.setByPayload);
-await brightness.init();
+subscribe(SeelenEvent.SystemMonitorsBrightnessChanged, brightness.setByPayload);
 
 let workspaces = lazyRune(() => invoke(SeelenCommand.StateGetVirtualDesktops));
-await subscribe(SeelenEvent.VirtualDesktopsChanged, workspaces.setByPayload);
-await workspaces.init();
+subscribe(SeelenEvent.VirtualDesktopsChanged, workspaces.setByPayload);
 
 let notifications = lazyRune(() => invoke(SeelenCommand.GetNotifications));
-await subscribe(SeelenEvent.Notifications, notifications.setByPayload);
-await notifications.init();
+subscribe(SeelenEvent.Notifications, notifications.setByPayload);
+
+await Promise.all([
+  mediaDevices.init(),
+  mediaPlaying.init(),
+  brightness.init(),
+  workspaces.init(),
+  notifications.init(),
+]);
 
 export const state = {
   get mediaInputs() {