+++ src/ui/svelte/keyboard-selector/state.svelte.ts
@@ -2,7 +2,7 @@ import { invoke, SeelenCommand, SeelenEvent, subscribe } from "@seelen-ui/lib";
 import { lazyRune } from "libs/ui/svelte/utils";
 
 const langs = lazyRune(() => invoke(SeelenCommand.SystemGetLanguages));
-await subscribe(SeelenEvent.SystemLanguagesChanged, langs.setByPayload);
+subscribe(SeelenEvent.SystemLanguagesChanged, langs.setByPayload);
 await langs.init();
 
 export const state = {